backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.3.5
pytest==7.4.3
transformers==4.35.2
torch==2.1.1
accelerate==0.24.1
bitsandbytes==0.41.1
safetensors==0.4.0
websockets==12.0
jinja2==3.1.2
aiofiles==23.2.1 